refactor utilities + enable non-multiprocess mode
